[3.0.0] Spire Admin (#109)
* Test

* Update Connections.vue

* Update Navbar.vue

* Add escape button back into quest api explorer

* Update Navbar.vue

* Fix Github login

* Suck less

* Break page and hotkeys

* Spire initialization work

* More init work

* More init work

* Finish user login

* Fix issues with connection page

* Tweaks to auth

* Auto inject local database instance into db connection

* Fix database connection update info logic

* Connection management work / cleanup

* Local user management

* Navbar admin fixes

* Password reset

* Change permissions response code to 403 instead of 401, be more descriptive about permission error return

* Implement Occulus proxy logic

* Get first Occulus page ported over

* More occulus work

* UI tweaks

* Port players online page

* UI tweaks

* Zone Servers

* Ported netstat page

* Port configuration

* Port tools

* Port client asset public downloading

* Tweaks

* Misc tweaks

* Implement process management

* Move configuration nav

* Consolidate navs

* Cleanup UI

* Update ServerProcessButtonComponent.vue

* Server rules cleanup, deep link rule searching

* Tweak and cleanup server rules page and zoneservers page

* Cleaning up components

* Tweaks

* Occulus require info route

* Only run Occulus integration locally

* AppEnv route checks

* Update PlayersOnlineComponent.vue

* Tweaks

* Generate player event model

* Generate API pipeline

* Update wire_gen.go

* Gen PlayerEventLogs relationships

* Player event log viewer work

* More player event log work

* More player event work

* PLAYER_EVENT.REZ_ACCEPTED

* Update PlayerEventLogs.vue

* LogSettings

* [Occulus] Stop timer, dev process launch, info-error banner fixes

* Port over log streaming

* Cleanup zone log streaming

* Update Navbar.vue

* Run process management for Occulus when web service is booted

* eqemuserverapi

* Update hello_world_cmd.go

* EqemuAdminClient -> OcculusClient

* Log ui

* Move controllers to their domain boundary

* Hook up eqemuserverapi controller

* Change response struct

* Reload log settings in game

* Finish log settings

* Discord webhooks

* UI tweaks

* Hot reload management page

* Choose static zones from launcher options

* Update eqemuserverapi_controller.go

* Implement database backups

* Update LogSettings.vue

* Update ServerRules.vue

* Server reload api

* Update LogSettings.vue

* Update LogSettings.vue

* Update QuestHotReload.vue

* Player event log settings

* Bulk loading

* More filters

* Update PlayerEventLogs.vue

* More work

* Explorer work

* Loading, raw expansion

* UI wrap

* Count endpoint for pagination

* Pagination

* Add count endpoint to code generation

* UI tweaks

* Trade event render

* More UI work

* Handin events

* Update PlayerEventDisplayComponent.vue

* Update occulus_controller.go

* Update eq-ui.css

* Tweaks

* Telnet client hardening

* Fetch Occulus binary from FS when release request fails

* Nil check

* Nil check since switching to pointer

* Update app-env.ts

* Avatar fix

* Connection active fix

* Default local

* Connection delete fix

* Reload on delete

* Connection tweaks

* CSS tweals

* UI tweaks

* Update EQSpellPreviewTable.vue

* Add whereJson support

* Filters

* Add event filters

* Tweaks

* Fix the rate limiter

* Add self built releases tab, tab selection component propogation

* Watch selection updates

* Fix zero state

* Add hack / killed events

* Add contains filters

* Update PlayerEventLogs.vue

* CSS tweaks

* Cleanup

* Server config controller work

* More nav work

* More nav work

* Update PlayerEventLogs.vue

* Misc tweaks

* Combine stderr / stdout

* Swap report check

* Server config work

* UCS

* Port zone server config

* Remove old config routes / views

* Update ServerConfig.vue

* Update DiscordCrashWebhook.vue

* Code gen

* Port message of the day

* Update Navbar.vue

* Wrap up Websocket proof of concept

* Websocket tweaking

* Rename eqemuserverapi to eqemuserver

* Move spireuser to spire

* websocketserver to websocket

* Websocket authentication, client

* Updater work

* Release installation

* Build streaming

* Build settings, progress streaming, cleaning

* Branch work

* Update eqemuserver_controller.go

* Finish server update

* Remove Occulus Server Code management

* Cross platform connection check

* Cross platform support work (Windows testing)

* Telnet hardening

* Update ServerReload.vue

* Dump hardware info for dashboard swap

* System endpoints

* Remove zone server dependency on Occulus (Spire native)

* Update ZoneServers.vue

* resource endpoints

* Header work

* Migrate dashboard counter stats to Spire

* Feed Server Processes card from admin header data feed

* Port dashboard

* Consolidate path management logic

* Port client file exports

* Port manual backups to Spire

* Cleanups

* Automate "days back" to last release, exclude notes found in previous release, exclude releases

* Add hotfix replacement back

* Coffee

* Fix changelog partial contains to exact

* More changelog generator consistency tweaks

* File log streaming

* Search filtering, watch timer etc

* Update FileLogs.vue

* Text formatting

* Add line buffers for rendering performance

* File deletion and filters

* Search all files

* Code complete

* Replace nav and old file logs

* Update SpireInitialize.vue

* Update FileLogs.vue

* Update FileLogs.vue

* Add localaddress address to config

* Preflight checks, http streaming client

* Cleanup

* Kbar

* Cleanup

* .

* Add autofocus to pages

* Update PlayerEventLogs.vue

* Admin permissions start

* Permission visual improvements

* Cleaning up bubbling up error messages

* More permission work

* More cleanup for permissions

* Tweak author formatting

* Keyboard shortcuts modal

* Add item creation event to renderer

* Detect when window blurs

* Remove duplicate groundspawn event renderer

* Handle task event renders

* Split money, task updates renders, server creation

* Trader purchase event

* Trader events

* Update FileLogs.vue

* More key work

* Update Navbar.vue

* Update Navbar.vue

* Bubble backup error

* Add image lazy loading to changlog, have changelog take up whole page width

* Min image height, changelog padding

* Fix state bug with players online

* Initial 3.0.0 release notes

* Update CHANGELOG.md

* Update lazy-image-load.ts

* Update CHANGELOG.md

* Fade in lazy image load

* Tweaks

* tweaks

* Update Home.vue

* Scroll to hash logic improve

* Cleanup database backups view

* Fix https://github.com/Akkadius/spire/issues/107

* Address https://github.com/Akkadius/spire/issues/106

* Fix https://github.com/Akkadius/spire/issues/104

* Fix https://github.com/Akkadius/spire/issues/92

* Address https://github.com/Akkadius/spire/issues/95

* Address https://github.com/Akkadius/spire/issues/91

* Add tab deep linking to item editor

* Add tab deep linking to spell editor

* Fix extradmgskill, skillmodtype zero value display

* Fix https://github.com/Akkadius/spire/issues/100

* Implement https://github.com/Akkadius/spire/issues/98 and cleanup state logic for QuestApiExplorer

* Fix https://github.com/Akkadius/spire/issues/83

* Update NpcSpellListEditor.vue

* Update changelog

* Update CHANGELOG.md

* Update clipboard.ts

+++ internal/models/character_item_recast.go
@@ -1,9 +1,9 @@
 package models
 
 type CharacterItemRecast struct {
-	ID          uint   `json:"id" gorm:"Column:id"`
-	RecastType  uint16 `json:"recast_type" gorm:"Column:recast_type"`
-	Timestamp   uint   `json:"timestamp" gorm:"Column:timestamp"`
+	ID          uint `json:"id" gorm:"Column:id"`
+	RecastType  uint `json:"recast_type" gorm:"Column:recast_type"`
+	Timestamp   uint `json:"timestamp" gorm:"Column:timestamp"`
 }
 
 func (CharacterItemRecast) TableName() string {
